refactor: msg 컴포넌트화
