+++ front/src/components/Msg/index.tsx
@@ -1,0 +1,1 @@
+export { default as Msg } from './Msg';
