+++ front/src/components/index.tsx
@@ -2,3 +2,4 @@ export { SnowGlobeCanvas } from './SnowGlobeCanvas';
 export { Button } from './Button';
 export { HeaderText } from './HeaderText';
 export { Prev } from './Prev';
+export { Msg } from './Msg';
